+++ src/index.js
@@ -1,5 +1,9 @@
 module.exports = {
-    MusixMatch: require('./musixmatch'),
-    AzLyrics: require('./azlyrics'),
-    LyricsCom: require('./lyrics.com'),
+    MusixMatch: require('./modules/musixmatch'),
+    AzLyrics: require('./modules/azlyrics'),
+    LyricsCom: require('./modules/lyrics.com'),
+    SongLyricsCom: require('./modules/songlyrics.com'),
+
+    /* Utility */
+    validateHeaders: require('./util/verify_useragent')
 }