+++ typings/index.d.ts
@@ -22,4 +22,14 @@ declare module '4lyrics' {
         fetchLyrics(url: string): Promise<string[]>;
         parseLyrics(html: string): string[];
     }
+
+    export class SongLyricsCom {
+        get useragent(): string;
+        fetchURLs(query: string): Promise<string[]>;
+        parseURLs(html: string): Array<string>;
+        fetchLyrics(url: string): Promise<string[]>;
+        parseLyrics(html: string): string[];
+    }
+
+    export function validateHeaders() : boolean;
 }